Hotfix/second

+++ app/src/main/java/com/capstoneC23PS274/segar/data/remote/response/ResultResponse.kt
@@ -17,7 +17,7 @@ data class ResultResponse(
 data class ResultData(
 
 	@field:SerializedName("score")
-	val score: Int? = null,
+	val score: Double? = null,
 
 	@field:SerializedName("image")
 	val image: String? = null,
